Add AnalyzeController
